Fix: updateRate for showing Cover had wrong type resulting in frequent recalculation of cover with little to no time for other calculations (by Moa)
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@6366 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Tactical/DisplayCover.cpp
@@ -429,7 +429,7 @@ BOOLEAN GridNoOnScreenAndAround( const INT32& sGridNo, const UINT8& ubRadius )
 
 void DisplayCover( const BOOLEAN& forceUpdate )
 {
-	if ( gGameExternalOptions.ubCoverDisplayUpdateWait == -1 )
+	if ( gGameExternalOptions.ubCoverDisplayUpdateWait == (UINT16)-1 )
 	{
 		return;
 	}
@@ -948,7 +948,7 @@ void RemoveMinesObjectsFromViewArea()
 
 void DisplayMines( const BOOLEAN& forceUpdate )
 {
-	if ( gGameExternalOptions.ubCoverDisplayUpdateWait == -1 )
+	if ( gGameExternalOptions.ubCoverDisplayUpdateWait == (UINT16)-1 )
 	{
 		return;
 	}
